Trait inexor_rgf_config_api::config_manager::ConfigManager

pub trait ConfigManager: Send + Sync + Lifecycle {

    // Required methods
    fn get_instance_config_location(&self) -> PathBuf;
    fn set_instance_config_location(&self, instance_config_location: PathBuf);
    fn get_graphql_server_config_location(&self) -> PathBuf;
    fn set_graphql_server_config_location(
        &self,
        graphql_server_config_location: PathBuf
    );
    fn get_plugins_config_location(&self) -> PathBuf;
    fn set_plugins_config_location(&self, plugins_config_location: PathBuf);
    fn get_remotes_config_location(&self) -> PathBuf;
    fn set_remotes_config_location(&self, remotes_config_location: PathBuf);
    fn get_instance_config(&self) -> InstanceConfig;
    fn set_instance_config(&self, instance_config: InstanceConfig);
    fn read_instance_config(&self);
    fn set_instance_name(&self, instance_name: &str);
    fn set_instance_description(&self, instance_description: &str);
    fn get_graphql_server_config(&self) -> GraphQLServerConfig;
    fn set_graphql_server_config(
        &self,
        graphql_server_config: GraphQLServerConfig
    );
    fn read_graphql_server_config(&self);
    fn set_graphql_hostname(&self, hostname: &str);
    fn set_graphql_port(&self, port: u16);
    fn set_graphql_secure(&self, secure: bool);
    fn set_graphql_ssl_certificate_path(&self, ssl_certificate_path: &str);
    fn set_graphql_ssl_private_key_path(&self, ssl_private_key_path: &str);
    fn set_graphql_shutdown_timeout(&self, shutdown_timeout: u64);
    fn set_graphql_workers(&self, workers: usize);
    fn get_graphql_default_context_path(&self) -> Option<String>;
    fn set_graphql_default_context_path(&self, default_context_path: String);
    fn get_plugins_config(&self) -> PluginsConfig;
    fn set_plugins_config(&self, plugins_config: PluginsConfig);
    fn read_plugins_config(&self);
    fn set_disable_all_plugins(&self, disable_all_plugins: bool);
    fn set_disabled_plugins(&self, disabled_plugins: Vec<String>);
    fn set_enabled_plugins(&self, enabled_plugins: Vec<String>);
    fn set_disable_hot_deploy(&self, disable_hot_deploy: bool);
    fn set_hot_deploy_location(&self, hot_deploy_location: Option<String>);
    fn set_install_location(&self, install_location: Option<String>);
    fn get_remotes_config(&self) -> RemotesConfig;
    fn set_remotes_config(&self, remotes: RemotesConfig);
    fn read_remotes_config(&self);
    fn write_remotes_config(&self);

}

Required Methods

fn get_instance_config_location(&self) -> PathBuf

Returns the location of the configuration of the instance.

fn set_instance_config_location(&self, instance_config_location: PathBuf)

Sets the location of the configuration of the instance.

fn get_graphql_server_config_location(&self) -> PathBuf

Returns the location of the configuration of the GraphQL server.

fn set_graphql_server_config_location( &self, graphql_server_config_location: PathBuf )

Sets the location of the configuration of the GraphQL server.

fn get_plugins_config_location(&self) -> PathBuf

Returns the location of the plugins configuration.

fn set_plugins_config_location(&self, plugins_config_location: PathBuf)

Sets the location of the plugins configuration.

fn get_remotes_config_location(&self) -> PathBuf

Returns the location of the remotes configuration.

fn set_remotes_config_location(&self, remotes_config_location: PathBuf)

Sets the location of the remotes configuration.

fn get_instance_config(&self) -> InstanceConfig

Returns the configuration of the instance.

fn set_instance_config(&self, instance_config: InstanceConfig)

Sets the configuration of the instance.

fn read_instance_config(&self)

Reads the configuration of the instance from file.

fn set_instance_name(&self, instance_name: &str)

Sets the instance name.

fn set_instance_description(&self, instance_description: &str)

Sets the instance description.

fn get_graphql_server_config(&self) -> GraphQLServerConfig

Returns the configuration of the GraphQL server.

fn set_graphql_server_config(&self, graphql_server_config: GraphQLServerConfig)

Sets the configuration of the GraphQL server.

fn read_graphql_server_config(&self)

Reads the configuration of the GraphQL server from file.

fn set_graphql_hostname(&self, hostname: &str)

Sets the host name.

fn set_graphql_port(&self, port: u16)

Sets the port.

fn set_graphql_secure(&self, secure: bool)

Enables / disables HTTPS.

fn set_graphql_ssl_certificate_path(&self, ssl_certificate_path: &str)

Sets the SSL certificate path.

fn set_graphql_ssl_private_key_path(&self, ssl_private_key_path: &str)

Sets the SSL private key path.

fn set_graphql_shutdown_timeout(&self, shutdown_timeout: u64)

Sets the timeout for graceful workers shutdown in seconds.

fn set_graphql_workers(&self, workers: usize)

Sets the number of workers.

fn get_graphql_default_context_path(&self) -> Option<String>

Returns the default context path which redirects the root context to a web resource provider.

fn set_graphql_default_context_path(&self, default_context_path: String)

Sets the default context path which redirects the root context to a web resource provider.

fn get_plugins_config(&self) -> PluginsConfig

Returns the plugins configuration.

fn set_plugins_config(&self, plugins_config: PluginsConfig)

Sets the plugins configuration.

fn read_plugins_config(&self)

Reads the plugins configuration from file.

fn set_disable_all_plugins(&self, disable_all_plugins: bool)

Enables / disables all plugins.

fn set_disabled_plugins(&self, disabled_plugins: Vec<String>)

Sets the plugins to disable.

fn set_enabled_plugins(&self, enabled_plugins: Vec<String>)

Sets the plugins to enable. If set, set_disabled_plugins will have no effect.

fn set_disable_hot_deploy(&self, disable_hot_deploy: bool)

Enables / disables hot deploy.

fn set_hot_deploy_location(&self, hot_deploy_location: Option<String>)

Sets the plugins hot deploy location.

fn set_install_location(&self, install_location: Option<String>)

Sets the plugins install location.

fn get_remotes_config(&self) -> RemotesConfig

Returns the remotes.

fn set_remotes_config(&self, remotes: RemotesConfig)

Sets the remotes configuration.

fn read_remotes_config(&self)

Reads the remotes configuration from file.

fn write_remotes_config(&self)

Writes the remotes configuration to file.